Complex operations optimized.

+++ Wearable/src/main/kotlin/zir/teq/wearable/watchface/util/DrawUtil.kt
@@ -136,7 +136,7 @@ class DrawUtil {
             val div = 1F + wave.lastWeight
             val newMagnitude = (complexPixel.magnitude + (wave.lastWeight * last.magnitude)) / div
             val newPhase = (complexPixel.phase + (wave.lastWeight * last.phase)) / div
-            val new = Complex.fromMagnitudeAndPhase(newMagnitude.toFloat(), newPhase)
+            val new = Complex.fromMagnitudeAndPhase(newMagnitude, newPhase)
 
             lastFrame.put(key, new)
             return ColorUtil.getColor(new)
